// src/types/store.interface.ts
export interface Store {
  id: number;
  name: string;
  title: string;
  seo_title: string;
  description: string;
  short_desc: string;
  message: string;
  data: string;
  img: string;
  alt: string;
  status: string;
  options: string;
  statu: string;
  link: string;
  slug: string;
  category_id: number;
  created_at: string;
  updated_at: string;
  deleted_at: string;
}